# Simplify the last-value aggregate