--- a/sdk/metric/internal/aggregate/aggregate.go
+++ b/sdk/metric/internal/aggregate/aggregate.go
@@ -44,7 +44,7 @@ type Builder[N int64 | float64] struct {
 
 func (b Builder[N]) filter(f Measure[N]) Measure[N] {
 	if b.Filter != nil {
-		fltr := b.Filter // Copy to prevent any funny stuff.
+		fltr := b.Filter // Copy to make it immutable after assignment.
 		return func(ctx context.Context, n N, a attribute.Set) {
 			fAttr, _ := a.Filter(fltr)
 			f(ctx, n, fAttr)